move MicroservicesRegistrar class to internal.deployment.msf4j package
