+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/deployment/msf4j/MicroservicesRegistrar.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal.http.msf4j;
+package org.wso2.carbon.uis.internal.deployment.msf4j;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceRegistration;
@@ -28,7 +28,6 @@ import org.osgi.service.component.annotations.ReferenceCardinality;
 import org.osgi.service.component.annotations.ReferencePolicy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.wso2.carbon.uis.internal.deployment.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.http.HttpTransport;
 import org.wso2.msf4j.Microservice;
 import org.wso2.msf4j.MicroservicesServer;
